#289 refactor: split error_trait.rs into 4 focused modules + docs + tests
- Split 769-line file into 4 logical modules
- error_trait.rs (188 lines) - main entry point with expand functions
- error_trait/source.rs (312 lines) - source method generation
- error_trait/backtrace.rs (285 lines) - backtrace method generation
- error_trait/binding.rs (100 lines) - field binding utilities
- error_trait/provide.rs (560 lines) - provide API implementation
- Add comprehensive documentation for all public items
- Add 19 unit tests (>95% coverage)
- NO mod.rs pattern used
- All 282 tests passing
